added option to disable tmp-server folder deletion

+++ server-factories/node-server-factory/src/main/kotlin/dev/redicloud/server/factory/ServerProcess.kt
@@ -172,10 +172,12 @@ class ServerProcess(
             }
         }
 
-        if (!configurationTemplate.static && !unexpectedlyStop) {
-            fileCopier.workDirectory.deleteRecursively()
-        }else if (unexpectedlyStop && !configurationTemplate.static) {
-            fileCopier.workDirectory.deleteOnExit()
+        if (System.getProperty("redicloud.server.delete-directory", "true").toBooleanStrictOrNull() == true) {
+            if (!configurationTemplate.static && !unexpectedlyStop) {
+                fileCopier.workDirectory.deleteRecursively()
+            }else if (unexpectedlyStop && !configurationTemplate.static) {
+                fileCopier.workDirectory.deleteOnExit()
+            }
         }
